+++ rule_runner_label.go
@@ -26,7 +26,6 @@ const (
 	compatMacOS260L
 	compatMacOS260XL
 	compatWindows2022
-	compatWindows2025
 	compatWindows2025VS2026
 	compatWindows11Arm
 	compatWindows11VS2026Arm
@@ -100,8 +99,8 @@ var defaultRunnerOSCompats = map[string]runnerOSCompat{
 	"xcode-27":               compatXcode270,
 	"xcode-27-xlarge":        compatXcode270XL,
 	"macos-latest-xlarge":    compatMacOS150XL,
-	"macos-latest-large":     compatMacOS150L,
-	"macos-latest":           compatMacOS150,
+	"macos-latest-large":     compatMacOS260L,
+	"macos-latest":           compatMacOS260,
 	"macos-26-intel":         compatMacOS260Intel,
 	"macos-26-xlarge":        compatMacOS260XL,
 	"macos-26-large":         compatMacOS260L,
@@ -113,16 +112,16 @@ var defaultRunnerOSCompats = map[string]runnerOSCompat{
 	"macos-14-xlarge":        compatMacOS140XL,
 	"macos-14-large":         compatMacOS140L,
 	"macos-14":               compatMacOS140,
-	"windows-latest":         compatWindows2022,
+	"windows-latest":         compatWindows2025VS2026,
 	"windows-latest-8-cores": compatWindows2022,
-	"windows-2025":           compatWindows2025,
+	"windows-2025":           compatWindows2025VS2026,
 	"windows-2025-vs2026":    compatWindows2025VS2026,
 	"windows-2022":           compatWindows2022,
 	"windows-11-arm":         compatWindows11Arm,
 	"windows-11-vs2026-arm":  compatWindows11VS2026Arm,
 	"linux":                  compatUbuntu2604 | compatUbuntu2404 | compatUbuntu2204, // Note: "linux" does not always indicate Ubuntu. It might be Fedora or Arch or ...
 	"macos":                  compatXcode270 | compatXcode270XL | compatMacOS260 | compatMacOS260Intel | compatMacOS260L | compatMacOS260XL | compatMacOS150 | compatMacOS150Intel | compatMacOS150L | compatMacOS150XL | compatMacOS140 | compatMacOS140L | compatMacOS140XL,
-	"windows":                compatWindows2025VS2026 | compatWindows2025 | compatWindows2022 | compatWindows11Arm | compatWindows11VS2026Arm,
+	"windows":                compatWindows2025VS2026 | compatWindows2022 | compatWindows11Arm | compatWindows11VS2026Arm,
 }
 
 // RuleRunnerLabel is a rule to check runner label like "ubuntu-latest". There are two types of